Django/requirements.txt
asgiref==3.8.1
Django==5.1.6
sqlparse==0.5.3
djangorestframework==3.12.4
